Improve Navigation Stack
